# Hack-Nation/docs/openapi.yaml
openapi: 3.1.0
info:
  title: PRISM inference API
  summary: Research-prototype PCOS evidence profiling. Not a diagnostic device.
  version: 0.1.0
paths:
...
  /api/v1/jobs/documents:
    post:
      tags:
      - jobs
      summary: Create Document Job
      operationId: create_document_job_api_v1_jobs_documents_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobSubmission'
        required: true
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobRecord'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/jobs/documents/{job_id}:
    get:
      tags:
      - jobs
      summary: Get Document Job
      operationId: get_document_job_api_v1_jobs_documents__job_id__get
      parameters:
      - name: job_id
        in: path
        required: true
        schema:
          type: string
          title: Job Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobRecord'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/jobs/speech:
    post:
      tags:
      - jobs
      summary: Create Speech Job
      operationId: create_speech_job_api_v1_jobs_speech_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobSubmission'
        required: true
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobRecord'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/jobs/speech/{job_id}:
    get:
      tags:
      - jobs
      summary: Get Speech Job
      operationId: get_speech_job_api_v1_jobs_speech__job_id__get
      parameters:
      - name: job_id
        in: path
        required: true
        schema:
          type: string
          title: Job Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobRecord'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/jobs/ultrasound:
    post:
      tags:
      - jobs
      summary: Create Ultrasound Job
      description: 'Accept an ultrasound study; report the inference gate as the reason.


        Uploads are accepted so the interface can be built and exercised. The job

        then terminates as ``unavailable`` carrying the same reason

        ``/models/status`` gives, so the UI never has to compose its own explanation

        for why an imaging result did not arrive.'
      operationId: create_ultrasound_job_api_v1_jobs_ultrasound_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobSubmission'
        required: true
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobRecord'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/jobs/ultrasound/{job_id}:
    get:
      tags:
      - jobs
      summary: Get Ultrasound Job
      operationId: get_ultrasound_job_api_v1_jobs_ultrasound__job_id__get
      parameters:
      - name: job_id
        in: path
        required: true
        schema:
          type: string
          title: Job Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobRecord'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    JobRecord:
      properties:
        job_id:
          type: string
          title: Job Id
        kind:
          type: string
          enum:
          - documents
          - speech
          - ultrasound
          title: Kind
        patient_id:
          type: string
          title: Patient Id
        status:
          type: string
          enum:
          - queued
          - processing
          - completed
          - failed
          - unavailable
          title: Status
        created_at:
          type: string
          title: Created At
        updated_at:
          type: string
          title: Updated At
        reason:
          anyOf:
          - type: string
          - type: 'null'
          title: Reason
        result:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Result
      additionalProperties: false
      type: object
      required:
      - job_id
      - kind
      - patient_id
      - status
      - created_at
      - updated_at
      title: JobRecord
      description: A job's observable state.
    JobSubmission:
      properties:
        patient_id:
          type: string
          minLength: 1
          title: Patient Id
        source_ids:
          items:
            type: string
          type: array
          title: Source Ids
        note:
          anyOf:
          - type: string
          - type: 'null'
          title: Note
      additionalProperties: false
      type: object
      required:
      - patient_id
      title: JobSubmission
      description: What a client sends to open a job.
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...